// src/types.ts
export enum Condition {
  RAIN = 'RAIN',
  SNOW = 'SNOW',
  SLEET = 'SLEET',
  FOG = 'FOG',
  WIND = 'WIND',
  CLEAR = 'CLEAR',
  PARTLY_CLOUDY = 'PARTLY_CLOUDY',
  CLOUDY = 'CLOUDY'
}

export enum UnitSystem {
  IMPERIAL = 'IMPERIAL',
  METRIC = 'METRIC',
  FREEDOM_UNITS = 'IMPERIAL',
  DEFAULT = 'DEFAULT'
}

export enum TempUnit {
  K = 'K',
  C = 'C',
  F = 'F'
}

export enum PressureUnit {
  MB = 'MB'
}

export enum VolumeUnit {
  MM = 'MM',
  IN = 'IN'
}

export enum VisibilityUnit {
  M = 'M',
  MI = 'MI',
  Percent = '%'
}

export enum WindUnit {
  MS = 'MS',
  MPH = 'MPH'
}